The d20 on the TurnOrder of each one is now from 1 to 20 and not 0 to 20 like before.
Corrections on some notes.

Correction on the enemy attack, now after the warrior dies the priest starts to take damage.

Correction on the print of the turn orders.

Putting the combat up to 4 enemies.

Damage changed to damage = WP from the attacker - AP from the receiver. Hp = Hp - damage.

Changes in the functions of the attacks and spells.

Order correction on the first turn.

Update on the end-game terms.