Update - 5
Some bug fixs.

New input “Press “enter”” after every action to give time to the player see the current action or what happen after them.

Setting a timer to show the enemy attack 1 by 1 to give time to the player see the enemy damage.

Clear the text before a new round.

Adding color to the text for a better view of what is happening.

Added more notes.

Fix on the turn order, now all the characters play in the round they must play.